requirements.txt
torch>=2.5.1
torchvision>=0.20.1
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0.74
tqdm>=4.66.0
pyyaml>=6.0.1
loralib>=0.1.2
einops>=0.7.0
omegaconf>=2.3.0
hydra-core>=1.3.2
# SAM2 will be installed from GitHub during Vast run to pick latest 2.1 code
vastai-sdk
